Add server for persistence of the deployed escrow contracts
Add an express server as a backend which persists the contract addresses
in a simple text file and has a function to retrieve the data.
Add the necessary changes to the front end to save and load the contract
addresses and load all other required data from the blockchain.

Note: These features aren't nearly production ready, it's missing better
error handling (currently mainly console logs) and validations, security
(anyone can call "/contractDeployed" on the server), etc.